Extract parseVintedItems pure function from the Vinted scanner
runVintedCheck mixed request/delay/SSE orchestration with ~135 lines of brittle
HTML/JSON item parsing (the catalog data-props blob, a bracket-matching "items"
regex, a feed-grid fallback, and a global href regex) — untestable without a
live HTTP flow, which is exactly why the price-group and object-leak bugs hid
there earlier.

Move all four parsing paths into a pure parseVintedItems(html, title, author):
VintedItem[] in services/vintedParser.ts. The scanner now calls it and keeps
only orchestration; runVintedCheck's parsing collapses to one line and
vintedSyncService drops 317 -> 184 lines.

Unit tests on captured HTML/JSON: catalog-blob parsing, relevance filtering,
the 5-item cap, no price-object leak on null amount, feed-grid fallback reading
the amount (not the whole aria-label), and empty on no match. 145 passing, lint
clean.
